# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "LumenMedia.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:15080/"
    }
  ],
  "paths": {
...
    "/api/v1/playback/decision": {
      "post": {
        "tags": [
          "Playback"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaybackDecisionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaybackDecisionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlaybackDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackDecisionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackDecisionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackDecisionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/playback/{sessionId}/set-quality": {
      "post": {
        "tags": [
          "Playback"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetQualityRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetQualityRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetQualityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackDecisionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackDecisionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackDecisionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/playback/{sessionId}/seek": {
      "post": {
        "tags": [
          "Playback"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeekRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SeekRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SeekRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackDecisionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackDecisionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackDecisionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/playback/{sessionId}/stop": {
      "post": {
        "tags": [
          "Playback"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/progress/{itemId}": {
      "put": {
        "tags": [
          "Progress"
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgressRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgressRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Progress"
        ],
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/continue-watching": {
      "get": {
        "tags": [
          "Progress"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfMediaItemSummary"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfMediaItemSummary"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfMediaItemSummary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtworkUrls": {
        "type": "object",
        "properties": {
          "poster": {
            "type": [
              "null",
              "string"
            ]
          },
          "backdrop": {
            "type": [
              "null",
              "string"
            ]
          },
          "logo": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumb": {
            "type": [
              "null",
              "string"
            ]
          },
          "banner": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "AudioStreamOption": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "language": {
            "type": [
              "null",
              "string"
            ]
          },
          "codec": {
            "type": [
              "null",
              "string"
            ]
          },
          "channels": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "isDefault": {
            "type": "boolean"
          }
        }
      },
...
      "DeviceProfile": {
        "type": "object",
        "properties": {
          "maxResolution": {
            "type": [
              "null",
              "string"
            ]
          },
          "maxBitrateKbps": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "videoCodecs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "audioCodecs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "containers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subtitleFormats": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "supportsHevc": {
            "type": "boolean"
          },
          "supportsHdr": {
            "type": "boolean"
          }
        }
      },
...
      "EpisodeSummary": {
        "required": [
          "id",
          "seriesId",
          "seasonId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/MediaKind"
          },
          "seriesId": {
            "type": "string",
            "format": "uuid"
          },
          "seasonId": {
            "type": "string",
            "format": "uuid"
          },
          "seasonNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "episodeNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "overview": {
            "type": [
              "null",
              "string"
            ]
          },
          "airDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "runtimeMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "artwork": {
            "$ref": "#/components/schemas/ArtworkUrls"
          },
          "userData": {
            "$ref": "#/components/schemas/UserDataDto"
          }
        }
      },
...
      "MediaItemSummary": {
        "required": [
          "id",
          "kind",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/MediaKind"
          },
          "title": {
            "type": "string"
          },
          "originalTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "year": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "runtimeMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "communityRating": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "officialRating": {
            "type": [
              "null",
              "string"
            ]
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "artwork": {
            "$ref": "#/components/schemas/ArtworkUrls"
          },
          "userData": {
            "$ref": "#/components/schemas/UserDataDto"
          },
          "addedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MediaKind": {
        "enum": [
          "Movie",
          "Series",
          "Episode"
        ],
        "type": "string"
      },
...
      "PagedResultOfMediaItemSummary": {
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaItemSummary"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PlaybackDecisionRequest": {
        "required": [
          "mediaId"
        ],
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "string",
            "format": "uuid"
          },
          "mediaSourceId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "mode": {
            "$ref": "#/components/schemas/PlaybackMode"
          },
          "qualityId": {
            "type": [
              "null",
              "string"
            ]
          },
          "audioStreamId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "subtitleStreamId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "resumePositionMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "profile": {
            "$ref": "#/components/schemas/DeviceProfile"
          }
        }
      },
      "PlaybackDecisionResponse": {
        "required": [
          "sessionId",
          "method",
          "mode",
          "streamUrl",
          "container",
          "selectedQualityId"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "method": {
            "$ref": "#/components/schemas/PlaybackMethod"
          },
          "mode": {
            "$ref": "#/components/schemas/PlaybackMode"
          },
          "streamUrl": {
            "type": "string"
          },
          "container": {
            "type": "string"
          },
          "startPositionMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "durationMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "selectedQualityId": {
            "type": "string"
          },
          "availableQualities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QualityOption"
            }
          },
          "audioStreams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AudioStreamOption"
            }
          },
          "subtitleStreams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubtitleStreamOption"
            }
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PlaybackMethod": {
        "enum": [
          "DirectPlay",
          "DirectStream",
          "Transcode"
        ],
        "type": "string"
      },
      "PlaybackMode": {
        "enum": [
          "auto",
          "manual"
        ],
        "type": "string"
      },
...
      "ProgressResponse": {
        "required": [
          "itemId"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "positionMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "watched": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "QualityOption": {
        "required": [
          "id",
          "label"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "adaptive": {
            "type": "boolean"
          },
          "width": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "height": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "bitrateKbps": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "SeekRequest": {
        "type": "object",
        "properties": {
          "positionMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "SetQualityRequest": {
        "required": [
          "qualityId"
        ],
        "type": "object",
        "properties": {
          "qualityId": {
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/PlaybackMode"
          },
          "resumePositionMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "audioStreamId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "subtitleStreamId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
      "SubtitleStreamOption": {
        "required": [
          "id",
          "deliveryUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "language": {
            "type": [
              "null",
              "string"
            ]
          },
          "format": {
            "type": [
              "null",
              "string"
            ]
          },
          "deliveryUrl": {
            "type": "string"
          }
        }
      },
...
      "UpdateProgressRequest": {
        "type": "object",
        "properties": {
          "positionMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "durationMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "sessionId": {
            "type": [
              "null",
              "string"
            ]
          },
          "state": {
            "type": "string"
          },
          "watched": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
...
      "UserDataDto": {
        "type": "object",
        "properties": {
          "watched": {
            "type": "boolean"
          },
          "playbackPositionMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "nextUp": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EpisodeSummary"
              }
            ]
          }
        }
      },
...
    }
  },
...
}